Script simplification, prepared to merge.
